MallocDeques.jl

MallocDeques.jl

MallocDeques.MallocDeques — Module

A StaticTools.jl and StaticCompilation.jl friendly double-ended queue implementation based on DataStructures.jl.

It implements only a small subset of DataStructures.jl's Deque.

julia> using MallocDeques
julia> a = MallocDeque{Int}()
MallocDeque with 0 elements.
julia> isempty(a)
true
julia> length(a)
0
julia> eltype(a)
Int64
julia> push!(a, 10)
MallocDeque with 1 elements.
julia> pop!(a)
10
julia> pushfirst!(a, 10)
MallocDeque with 1 elements.
julia> push!(a, 20)
MallocDeque with 2 elements.
julia> first(a)
10
julia> last(a)
20
julia> popfirst!(a)
10
julia> push!(a, 10)
MallocDeque with 1 elements.
julia> empty!(a)
MallocDeque with 0 elements.

StaticCompiler.jl

In order to static compilation to work, no exception can be thrown. That means that all boundary check were removed, so pop!-ing an empty deque will result in unpredicted behavior.

MallocDeque also implements StaticTools.free.

MallocDeques.MallocDeque — Type

MallocDeque{T, C}

Constructs MallocDeque object for elements of type T. Parameters ––––– T::Type MallocDeque element data type. C::Int MallocDeque block size (in bytes). Default = 1024.

Base.empty! — Method

empty!(d::MallocDeque)

Reset the deque d.

Base.first — Method

first(d::MallocDeque)

Returns the first element of the deque d.

Base.isempty — Method

isempty(d::MallocDeque)

Verifies if deque d is empty.

Base.last — Method

last(d::MallocDeque)

Returns the last element of the deque d.

Base.length — Method

length(d::MallocDeque)

Returns the number of elements in deque d.

Base.pop! — Method

pop!(d::MallocDeque{T,C}) where {T,C}

Remove the element at the back of deque d. Since throwing errors is not allowed, poping an empty deque will result in pain and suffering.

Base.popfirst! — Method

popfirst!(d::MallocDeque{T,C}) where {T,C}

Remove the element at the front of deque d. Since throwing errors is not allowed, poping an empty deque will result in pain and suffering.

Base.push! — Method

push!(d::MallocDeque{T,C}, x::T) where {T,C}

Add an element to the back of deque d.

Base.pushfirst! — Method

pushfirst!(d::MallocDeque{T,C}, x::T) where {T,C}

Add an element to the front of deque d.